requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.6.3
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.13
langsmith==0.1.83
marshmallow==3.20.2
multidict==6.0.4
mwoauth==0.4.0
mypy-extensions==1.0.0
numpy==1.24.4
oauthlib==3.2.2
openai==1.35.8
packaging==23.2
pydantic==2.5.3
pydantic-core==2.14.6
PyJWT==2.8.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.0
SQLAlchemy==2.0.25
tenacity==8.2.3
tiktoken==0.7.0
tqdm==4.66.1
typing-extensions==4.9.0
typing-inspect==0.9.0
ujson==5.9.0
urllib3==2.1.0
wikibaseintegrator==0.12.5
yarl==1.9.4
wikibase-rest-api-client==0.1.3
mediawikiapi==1.2
google-generativeai==0.7.1
langchain-google-genai==1.0.7
gradio==4.22.0
python-dotenv==1.0.1
